+++ plugins/org.chromium.sdk/src/org/chromium/sdk/internal/v8native/value/ValueLoader.java
@@ -4,11 +4,11 @@
 
 package org.chromium.sdk.internal.v8native.value;
 
-import java.util.List;
-
 import org.chromium.sdk.RemoteValueMapping;
 import org.chromium.sdk.internal.v8native.InternalContext;
 import org.chromium.sdk.util.MethodIsBlockingException;
+
+import java.util.List;
 
 /**
  * The elaborate factory and storage for {@link ValueMirror}'s, that loads values from remote and
@@ -28,16 +28,14 @@ public abstract class ValueLoader implements RemoteValueMapping {
    * Looks up {@link ValueMirror} in map, loads them if needed or reloads them
    * if property data is unavailable (or expired).
    */
-  public abstract SubpropertiesMirror getOrLoadSubproperties(long ref)
-      throws MethodIsBlockingException;
+  public abstract SubpropertiesMirror getOrLoadSubproperties(int ref) throws MethodIsBlockingException;
 
   /**
    * For each PropertyReference from propertyRefs tries to either:
    * 1. read it from PropertyReference (possibly cached value) or
    * 2. lookup value by refId from remote
    */
-  public abstract ValueMirror[] getOrLoadValueFromRefs(
-      List<? extends PropertyReference> propertyRefs) throws MethodIsBlockingException;
+  public abstract ValueMirror[] getOrLoadValueFromRefs(List<? extends PropertyReference> propertyRefs) throws MethodIsBlockingException;
 
   public abstract InternalContext getInternalContext();
 }
